setup for eventually updating leaderboards; further testing

+++ src/commands/guess.ts
@@ -1,6 +1,9 @@
-import { CommandInteraction, InteractionReplyOptions, Message, SlashCommandBuilder, User } from "discord.js";
+import { config } from "../config";
+import { CommandInteraction, InteractionReplyOptions, SlashCommandBuilder, User } from "discord.js";
 import { Question, QuestionPost } from "../types";
 import { setQuestionAnswered } from "..";
+
+const allowedChannelId: string = config.CHANNEL_ID;
 
 export const data = new SlashCommandBuilder()
   .setName("guess")
@@ -12,6 +15,11 @@ export const data = new SlashCommandBuilder()
 
 export async function execute(interaction: CommandInteraction, questionPost: QuestionPost, currentQuestionAnswered: boolean) {
   try {
+    if (interaction.channelId != allowedChannelId) {
+      await interaction.reply({content: "Wrong channel!", ephemeral: true});
+      return;
+    }
+
     const answerOption = interaction.options.get("answer");
     const user: User = interaction.user;
 
@@ -27,7 +35,7 @@ export async function execute(interaction: CommandInteraction, questionPost: Que
         await interaction.reply("The current question was already answered! Wait for the next one.");
       } else {
         if (answer == question.answer) {
-          await handleCorrectAnswer(interaction, user, question, questionPost);
+          await handleCorrectAnswer(interaction, user, questionPost);
         } else {
           await interaction.reply(question.wrongResponse);
         }
@@ -38,12 +46,17 @@ export async function execute(interaction: CommandInteraction, questionPost: Que
   }
 }
 
-async function handleCorrectAnswer(interaction: CommandInteraction, user: User, question: Question, questionPost: QuestionPost) {
+async function handleCorrectAnswer(interaction: CommandInteraction, user: User, questionPost: QuestionPost) {
   try {
+    if (!questionPost.question) {
+      throw new Error("Question is undefined");
+    }
     setQuestionAnswered();
-    const reply: string = user.username + " " + question.correctResponse;
+    const reply: string = user.username + " " + questionPost?.question.correctResponse;
     const replyOptions: InteractionReplyOptions = { content: reply };
     await interaction.reply(replyOptions);
+
+    await updateExpAndDrachma(user.id, questionPost.question);
 
     setTimeout(async () => {
       await interaction.deleteReply();
@@ -54,4 +67,10 @@ async function handleCorrectAnswer(interaction: CommandInteraction, user: User, 
   } catch (error) {
     console.error("Error handling correct answer:", error);
   }
+}
+
+async function updateExpAndDrachma(userId: string, question: Question) {
+  let drachma: number = Number(question.drachma.$numberInt);
+  let exp: number = Number(question.exp.$numberInt);
+  console.log(userId + " guessed correctly; updates: drachma +" + drachma + "; exp +" + exp); // TODO: update to a proper database for this
 }
